# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Daily Management Express API",
    "description": "Daily Management REST API documentation for mobile client and server integration.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3005",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter the Better Auth session token"
      }
    }
  },
  "paths": {
...
    "/api/daily-logs": {
      "get": {
        "summary": "Get daily log by date",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "date", "in": "query", "required": true, "schema": { "type": "string" }, "example": "2026-07-21" }
        ],
        "responses": {
          "200": { "description": "Returns daily log details" }
        }
      }
    },
    "/api/daily-logs/meals": {
      "post": {
        "summary": "Add a food entry to a daily log meal",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": { "type": "string", "example": "2026-07-21" },
                  "type": { "type": "string", "enum": ["breakfast", "lunch", "dinner", "snack"], "example": "breakfast" },
                  "food_name": { "type": "string", "example": "Egg" },
                  "serving_description": { "type": "string", "example": "1 adet" },
                  "quantity": { "type": "number", "example": 2 },
                  "calories": { "type": "number", "example": 140 },
                  "protein_g": { "type": "number", "example": 12 },
                  "carbs_g": { "type": "number", "example": 0 },
                  "fat_g": { "type": "number", "example": 10 },
                  "fatsecret_food_id": { "type": "string" },
                  "save_as_recipe": { "type": "boolean", "default": false }
                },
                "required": ["date", "type", "food_name", "serving_description", "quantity", "calories", "protein_g", "carbs_g", "fat_g"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Meal added successfully" }
        }
      },
      "put": {
        "summary": "Update a meal entry in a daily log",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": { "type": "string", "example": "2026-07-21" },
                  "entry_id": { "type": "string" },
                  "type": { "type": "string", "enum": ["breakfast", "lunch", "dinner", "snack"] },
                  "old_type": { "type": "string", "enum": ["breakfast", "lunch", "dinner", "snack"] },
                  "food_name": { "type": "string" },
                  "serving_description": { "type": "string" },
                  "quantity": { "type": "number" },
                  "calories": { "type": "number" },
                  "protein_g": { "type": "number" },
                  "carbs_g": { "type": "number" },
                  "fat_g": { "type": "number" }
                },
                "required": ["date", "entry_id", "type", "old_type", "food_name", "serving_description", "quantity", "calories", "protein_g", "carbs_g", "fat_g"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Meal entry updated successfully" }
        }
      }
    },
    "/api/daily-logs/meals/delete": {
      "post": {
        "summary": "Delete a food entry from a daily log meal",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": { "type": "string", "example": "2026-07-21" },
                  "entry_id": { "type": "string" },
                  "type": { "type": "string", "enum": ["breakfast", "lunch", "dinner", "snack"], "example": "breakfast" }
                },
                "required": ["date", "entry_id", "type"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Meal entry deleted successfully" }
        }
      }
    },
    "/api/daily-logs/sleep": {
      "post": {
        "summary": "Update daily log sleep information",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": { "type": "string", "example": "2026-07-21" },
                  "duration_minutes": { "type": "integer", "example": 480 },
                  "calories_burned": { "type": "number", "example": 50 },
                  "quality": { "type": "string", "example": "Good" }
                },
                "required": ["date", "duration_minutes", "calories_burned"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Sleep data updated successfully" }
        }
      }
    },
    "/api/daily-logs/exercises": {
      "post": {
        "summary": "Add an exercise entry to a daily log",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": { "type": "string", "example": "2026-07-21" },
                  "name": { "type": "string", "example": "Running" },
                  "duration_minutes": { "type": "integer", "example": 30 },
                  "calories_burned": { "type": "number", "example": 300 },
                  "source": { "type": "string", "enum": ["manual", "estimated"], "example": "manual" }
                },
                "required": ["date", "name", "duration_minutes", "calories_burned", "source"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Exercise entry added successfully" }
        }
      }
    },
    "/api/daily-logs/exercises/delete": {
      "post": {
        "summary": "Delete an exercise entry from a daily log",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": { "type": "string", "example": "2026-07-21" },
                  "entry_id": { "type": "string" }
                },
                "required": ["date", "entry_id"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Exercise entry deleted successfully" }
        }
      }
    },
...
  }
}